Various fixes and improvements. Replaced SqlRecordVector with typedef. Better debug output for flight plan. Added more methods for insert in DataManagerBase. Corrected return value for hash function for routing::Node

+++ src/routing/routenetworktypes.h
@@ -229,9 +229,9 @@ struct Node
 
 };
 
-inline int qHash(const atools::routing::Node& node)
+inline uint qHash(const atools::routing::Node& node)
 {
-  return node.index;
+  return static_cast<uint>(node.index);
 }
 
 struct Result
